successfully cleaned up file structure
