+++ client/src/components/SnapshotView.tsx
