+++ client/src/slices/snapshotSlice.ts
